data-indexing/2-data-ingestion/pyproject.toml
[project]
name = "2-data-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-servicebus>=7.14.2",
    "dotenv>=0.9.9",
    "openai>=1.82.1",
    "pika>=1.3.2",
    "pymongo>=4.12.0",
]
